refactor: Rename JAXBIO to JaxbSerializer
Class has also been moved to util package to avoid a single class package (io).
